Refactor serverRequestParams to perform related functions, functions relocated to related files
