+++ lib/msal-core/src/ServerRequestParameters.ts
@@ -3,6 +3,11 @@
 
 import { Authority } from "./Authority";
 import { Utils } from "./Utils";
+import { QPDict, AuthenticationParameters, validateClaimsRequest } from "./AuthenticationParameters";
+import { Account } from "./Account";
+import { SSOTypes, Constants, PromptState, BlacklistedEQParams } from "./Constants";
+import { ClientConfigurationError } from "./error/ClientConfigurationError";
+import { Logger } from "./Logger";
 
 /**
  * Nonce: OIDC Nonce definition: https://openid.net/specs/openid-connect-core-1_0.html#IDToken
@@ -45,7 +50,7 @@ export class ServerRequestParameters {
    * @param redirectUri
    * @param state
    */
-  constructor (authority: Authority, clientId: string, scope: Array<string>, responseType: string, redirectUri: string, state: string ) {
+  constructor (authority: Authority, clientId: string, scope: Array<string>, responseType: string, redirectUri: string, state: string) {
     this.authorityInstance = authority;
     this.clientId = clientId;
     this.scopes = scope;
@@ -65,98 +70,287 @@ export class ServerRequestParameters {
   }
 
   /**
-   * generates the URL with QueryString Parameters
-   * @param scopes
+   * @hidden
+   * @ignore
+   *
+   * Utility to populate QueryParameters and ExtraQueryParameters to ServerRequestParamerers
+   * @param request
+   * @param serverAuthenticationRequest
    */
-  createNavigateUrl(scopes: Array<string>): string {
-    const str = this.createNavigationUrlString(scopes);
-    let authEndpoint: string = this.authorityInstance.AuthorizationEndpoint;
-    // if the endpoint already has queryparams, lets add to it, otherwise add the first one
-    if (authEndpoint.indexOf("?") < 0) {
-      authEndpoint += "?";
-    } else {
-      authEndpoint += "&";
+  populateQueryParams(account: Account, request: AuthenticationParameters, adalIdTokenObject?: any): void {
+    let queryParameters: QPDict = {};
+
+    if (request) {
+      // add the prompt parameter to serverRequestParameters if passed
+      if (request.prompt) {
+        this.validatePromptParameter(request.prompt);
+        this.promptValue = request.prompt;
+      }
+
+      // Add claims challenge to serverRequestParameters if passed
+      if (request.claimsRequest) {
+        validateClaimsRequest(request);
+        this.claimsValue = request.claimsRequest;
+      }
+
+      // if the developer provides one of these, give preference to developer choice
+      if (ServerRequestParameters.isSSOParam(request)) {
+        queryParameters = this.constructUnifiedCacheQueryParameter(request, null);
+      }
     }
 
-    const requestUrl: string = `${authEndpoint}${str.join("&")}`;
-    return requestUrl;
+    if (adalIdTokenObject) {
+      queryParameters = this.constructUnifiedCacheQueryParameter(null, adalIdTokenObject);
+    }
+
+    // adds sid/login_hint if not populated; populates domain_req, login_req and domain_hint
+    // this.logger.verbose("Calling addHint parameters");
+    queryParameters = this.addHintParameters(account, queryParameters);
+
+    // sanity check for developer passed extraQueryParameters
+    let eQParams: QPDict;
+    if (request) {
+      eQParams = this.sanitizeEQParams(request);
+    }
+
+    // Populate the extraQueryParameters to be sent to the server
+    this.queryParameters = this.generateQueryParametersString(queryParameters);
+    this.extraQueryParameters = this.generateQueryParametersString(eQParams);
   }
 
-  /**
-   * Generate the array of all QueryStringParams to be sent to the server
-   * @param scopes
-   */
-  createNavigationUrlString(scopes: Array<string>): Array<string> {
-    if (!scopes) {
-      scopes = [this.clientId];
-    }
-
-    if (scopes.indexOf(this.clientId) === -1) {
-      scopes.push(this.clientId);
-    }
-    const str: Array<string> = [];
-    str.push("response_type=" + this.responseType);
-
-    this.translateclientIdUsedInScope(scopes);
-    str.push("scope=" + encodeURIComponent(this.parseScope(scopes)));
-    str.push("client_id=" + encodeURIComponent(this.clientId));
-    str.push("redirect_uri=" + encodeURIComponent(this.redirectUri));
-
-    str.push("state=" + encodeURIComponent(this.state));
-    str.push("nonce=" + encodeURIComponent(this.nonce));
-
-    str.push("client_info=1");
-    str.push(`x-client-SKU=${this.xClientSku}`);
-    str.push(`x-client-Ver=${this.xClientVer}`);
-    if (this.promptValue) {
-      str.push("prompt=" + encodeURIComponent(this.promptValue));
-    }
-
-    if (this.claimsValue) {
-      str.push("claims=" + encodeURIComponent(this.claimsValue));
-    }
-
-    if (this.queryParameters) {
-      str.push(this.queryParameters);
-    }
-
-    if (this.extraQueryParameters) {
-      str.push(this.extraQueryParameters);
-    }
-
-    str.push("client-request-id=" + encodeURIComponent(this.correlationId));
-    return str;
-  }
+  //#region QueryParam helpers
 
   /**
-   * append the required scopes: https://openid.net/specs/openid-connect-basic-1_0.html#Scopes
-   * @param scopes
+   * @hidden
+   * @ignore
+   *
+   * Utility to test if valid prompt value is passed in the request
+   * @param request
    */
-  translateclientIdUsedInScope(scopes: Array<string>): void {
-    const clientIdIndex: number = scopes.indexOf(this.clientId);
-    if (clientIdIndex >= 0) {
-      scopes.splice(clientIdIndex, 1);
-      if (scopes.indexOf("openid") === -1) {
-        scopes.push("openid");
-      }
-      if (scopes.indexOf("profile") === -1) {
-        scopes.push("profile");
-      }
+  private validatePromptParameter (prompt: string) {
+    if (!([PromptState.LOGIN, PromptState.SELECT_ACCOUNT, PromptState.CONSENT, PromptState.NONE].indexOf(prompt) >= 0)) {
+        throw ClientConfigurationError.createInvalidPromptError(prompt);
     }
   }
 
   /**
-   * Parse the scopes into a formatted scopeList
-   * @param scopes
+   * Constructs extraQueryParameters to be sent to the server for the AuthenticationParameters set by the developer
+   * in any login() or acquireToken() calls
+   * @param idTokenObject
+   * @param extraQueryParameters
+   * @param sid
+   * @param loginHint
    */
-  parseScope(scopes: Array<string>): string {
-    let scopeList: string = "";
-    if (scopes) {
-        for (let i: number = 0; i < scopes.length; ++i) {
-        scopeList += (i !== scopes.length - 1) ? scopes[i] + " " : scopes[i];
+  //TODO: check how this behaves when domain_hint only is sent in extraparameters and idToken has no upn.
+  private constructUnifiedCacheQueryParameter(request: AuthenticationParameters, idTokenObject: any): QPDict {
+
+    // preference order: account > sid > login_hint
+    let ssoType;
+    let ssoData;
+    let serverReqParam: QPDict = {};
+    // if account info is passed, account.sid > account.login_hint
+    if (request) {
+      if (request.account) {
+        const account: Account = request.account;
+        if (account.sid) {
+          ssoType = SSOTypes.SID;
+          ssoData = account.sid;
+        }
+        else if (account.userName) {
+          ssoType = SSOTypes.LOGIN_HINT;
+          ssoData = account.userName;
+        }
+      }
+      // sid from request
+      else if (request.sid) {
+        ssoType = SSOTypes.SID;
+        ssoData = request.sid;
+      }
+      // loginHint from request
+      else if (request.loginHint) {
+        ssoType = SSOTypes.LOGIN_HINT;
+        ssoData = request.loginHint;
+      }
+    }
+    // adalIdToken retrieved from cache
+    else if (idTokenObject) {
+      if (idTokenObject.hasOwnProperty(Constants.upn)) {
+        ssoType = SSOTypes.ID_TOKEN;
+        ssoData = idTokenObject.upn;
+      }
+      else {
+        ssoType = SSOTypes.ORGANIZATIONS;
+        ssoData = null;
       }
     }
 
-    return scopeList;
+    serverReqParam = this.addSSOParameter(ssoType, ssoData);
+
+    // add the HomeAccountIdentifier info/ domain_hint
+    if (request && request.account && request.account.homeAccountIdentifier) {
+        serverReqParam = this.addSSOParameter(SSOTypes.HOMEACCOUNT_ID, request.account.homeAccountIdentifier, serverReqParam);
+    }
+
+    return serverReqParam;
+  }
+
+  /**
+   * @hidden
+   *
+   * Adds login_hint to authorization URL which is used to pre-fill the username field of sign in page for the user if known ahead of time
+   * domain_hint can be one of users/organizations which when added skips the email based discovery process of the user
+   * domain_req utid received as part of the clientInfo
+   * login_req uid received as part of clientInfo
+   * Also does a sanity check for extraQueryParameters passed by the user to ensure no repeat queryParameters
+   *
+   * @param {@link Account} account - Account for which the token is requested
+   * @param queryparams
+   * @param {@link ServerRequestParameters}
+   * @ignore
+   */
+  private addHintParameters(account: Account, qParams: QPDict): QPDict {
+    // This is a final check for all queryParams added so far; preference order: sid > login_hint
+    // sid cannot be passed along with login_hint or domain_hint, hence we check both are not populated yet in queryParameters
+    if (account && !qParams[SSOTypes.SID]) {
+      // sid - populate only if login_hint is not already populated and the account has sid
+      const populateSID = !qParams[SSOTypes.LOGIN_HINT] && account.sid && this.promptValue === PromptState.NONE;
+      if (populateSID) {
+          qParams = this.addSSOParameter(SSOTypes.SID, account.sid, qParams);
+      }
+      // login_hint - account.userName
+      else {
+        const populateLoginHint = !qParams[SSOTypes.LOGIN_HINT] && account.userName && !Utils.isEmpty(account.userName);
+        if (populateLoginHint) {
+          qParams = this.addSSOParameter(SSOTypes.LOGIN_HINT, account.userName, qParams);
+        }
+      }
+
+      const populateReqParams = !qParams[SSOTypes.DOMAIN_REQ] && !qParams[SSOTypes.LOGIN_REQ];
+      if (populateReqParams) {
+        qParams = this.addSSOParameter(SSOTypes.HOMEACCOUNT_ID, account.homeAccountIdentifier, qParams);
+      }
+    }
+
+    return qParams;
+  }
+
+  /**
+   * Add SID to extraQueryParameters
+   * @param sid
+   */
+  private addSSOParameter(ssoType: string, ssoData: string, ssoParam?: QPDict): QPDict {
+    if (!ssoParam) {
+      ssoParam = {};
+    }
+
+    if (!ssoData) {
+        return ssoParam;
+    }
+
+    switch (ssoType) {
+      case SSOTypes.SID: {
+        ssoParam[SSOTypes.SID] = ssoData;
+        break;
+      }
+      case SSOTypes.ID_TOKEN: {
+        ssoParam[SSOTypes.LOGIN_HINT] = ssoData;
+        ssoParam[SSOTypes.DOMAIN_HINT] = SSOTypes.ORGANIZATIONS;
+        break;
+      }
+      case SSOTypes.LOGIN_HINT: {
+        ssoParam[SSOTypes.LOGIN_HINT] = ssoData;
+        break;
+      }
+      case SSOTypes.ORGANIZATIONS: {
+        ssoParam[SSOTypes.DOMAIN_HINT] = SSOTypes.ORGANIZATIONS;
+        break;
+      }
+      case SSOTypes.CONSUMERS: {
+        ssoParam[SSOTypes.DOMAIN_HINT] = SSOTypes.CONSUMERS;
+        break;
+      }
+      case SSOTypes.HOMEACCOUNT_ID: {
+        let homeAccountId = ssoData.split(".");
+        const uid = Utils.base64DecodeStringUrlSafe(homeAccountId[0]);
+        const utid = Utils.base64DecodeStringUrlSafe(homeAccountId[1]);
+
+        // TODO: domain_req and login_req are not needed according to eSTS team
+        ssoParam[SSOTypes.LOGIN_REQ] = uid;
+        ssoParam[SSOTypes.DOMAIN_REQ] = utid;
+
+        if (utid === Constants.consumersUtid) {
+            ssoParam[SSOTypes.DOMAIN_HINT] = SSOTypes.CONSUMERS;
+        }
+        else {
+            ssoParam[SSOTypes.DOMAIN_HINT] = SSOTypes.ORGANIZATIONS;
+        }
+        break;
+      }
+      case SSOTypes.LOGIN_REQ: {
+        ssoParam[SSOTypes.LOGIN_REQ] = ssoData;
+        break;
+      }
+      case SSOTypes.DOMAIN_REQ: {
+        ssoParam[SSOTypes.DOMAIN_REQ] = ssoData;
+        break;
+      }
+    }
+
+    return ssoParam;
+  }
+
+  /**
+   * @hidden
+   * @ignore
+   * Removes unnecessary or duplicate query parameters from extraQueryParameters
+   * @param request
+   */
+  private sanitizeEQParams(request: AuthenticationParameters) : QPDict {
+    let eQParams : QPDict = request.extraQueryParameters;
+    if (!eQParams) {
+      return null;
+    }
+    if (request.claimsRequest) {
+      // this.logger.warning("Removed duplicate claims from extraQueryParameters. Please use either the claimsRequest field OR pass as extraQueryParameter - not both.");
+      delete eQParams[Constants.claims];
+    }
+    BlacklistedEQParams.forEach(param => {
+      if (eQParams[param]) {
+        // this.logger.warning("Removed duplicate " + param + " from extraQueryParameters. Please use the " + param + " field in request object.");
+        delete eQParams[param];
+      }
+    });
+    return eQParams;
+  }
+
+  /**
+   * Utility to generate a QueryParameterString from a Key-Value mapping of extraQueryParameters passed
+   * @param extraQueryParameters
+   */
+  private generateQueryParametersString(queryParameters: QPDict): string {
+    let paramsString: string = null;
+
+    if (queryParameters) {
+      Object.keys(queryParameters).forEach((key: string) => {
+        if (paramsString == null) {
+          paramsString = `${key}=${encodeURIComponent(queryParameters[key])}`;
+        }
+        else {
+          paramsString += `&${key}=${encodeURIComponent(queryParameters[key])}`;
+        }
+     });
+    }
+
+    return paramsString;
+  }
+
+  //#endregion
+
+  /**
+   * Check to see if there are SSO params set in the Request
+   * @param request
+   */
+  static isSSOParam(request: AuthenticationParameters) {
+      return request && (request.account || request.sid || request.loginHint);
   }
 }
