+++ lib/msal-core/src/Utils.ts
@@ -322,164 +322,4 @@ export class Utils {
 
   //#endregion
 
-  //#region ExtraQueryParameters Processing (Extract?)
-
-  /**
-   * Constructs extraQueryParameters to be sent to the server for the AuthenticationParameters set by the developer
-   * in any login() or acquireToken() calls
-   * @param idTokenObject
-   * @param extraQueryParameters
-   * @param sid
-   * @param loginHint
-   */
-  //TODO: check how this behaves when domain_hint only is sent in extraparameters and idToken has no upn.
-  static constructUnifiedCacheQueryParameter(request: AuthenticationParameters, idTokenObject: any): QPDict {
-
-    // preference order: account > sid > login_hint
-    let ssoType;
-    let ssoData;
-    let serverReqParam: QPDict = {};
-    // if account info is passed, account.sid > account.login_hint
-    if (request) {
-      if (request.account) {
-        const account: Account = request.account;
-        if (account.sid) {
-          ssoType = SSOTypes.SID;
-          ssoData = account.sid;
-        }
-        else if (account.userName) {
-          ssoType = SSOTypes.LOGIN_HINT;
-          ssoData = account.userName;
-        }
-      }
-      // sid from request
-      else if (request.sid) {
-        ssoType = SSOTypes.SID;
-        ssoData = request.sid;
-      }
-      // loginHint from request
-      else if (request.loginHint) {
-        ssoType = SSOTypes.LOGIN_HINT;
-        ssoData = request.loginHint;
-      }
-    }
-    // adalIdToken retrieved from cache
-    else if (idTokenObject) {
-      if (idTokenObject.hasOwnProperty(Constants.upn)) {
-        ssoType = SSOTypes.ID_TOKEN;
-        ssoData = idTokenObject.upn;
-      }
-      else {
-        ssoType = SSOTypes.ORGANIZATIONS;
-        ssoData = null;
-      }
-    }
-
-    serverReqParam = this.addSSOParameter(ssoType, ssoData);
-
-    // add the HomeAccountIdentifier info/ domain_hint
-    if (request && request.account && request.account.homeAccountIdentifier) {
-        serverReqParam = this.addSSOParameter(SSOTypes.HOMEACCOUNT_ID, request.account.homeAccountIdentifier, serverReqParam);
-    }
-
-    return serverReqParam;
-  }
-
-
-  /**
-   * Add SID to extraQueryParameters
-   * @param sid
-   */
-  static addSSOParameter(ssoType: string, ssoData: string, ssoParam?: QPDict): QPDict {
-    if (!ssoParam) {
-      ssoParam = {};
-    }
-
-    if (!ssoData) {
-        return ssoParam;
-    }
-
-    switch (ssoType) {
-      case SSOTypes.SID: {
-        ssoParam[SSOTypes.SID] = ssoData;
-        break;
-      }
-      case SSOTypes.ID_TOKEN: {
-        ssoParam[SSOTypes.LOGIN_HINT] = ssoData;
-        ssoParam[SSOTypes.DOMAIN_HINT] = SSOTypes.ORGANIZATIONS;
-        break;
-      }
-      case SSOTypes.LOGIN_HINT: {
-        ssoParam[SSOTypes.LOGIN_HINT] = ssoData;
-        break;
-      }
-      case SSOTypes.ORGANIZATIONS: {
-        ssoParam[SSOTypes.DOMAIN_HINT] = SSOTypes.ORGANIZATIONS;
-        break;
-      }
-      case SSOTypes.CONSUMERS: {
-        ssoParam[SSOTypes.DOMAIN_HINT] = SSOTypes.CONSUMERS;
-        break;
-      }
-      case SSOTypes.HOMEACCOUNT_ID: {
-        let homeAccountId = ssoData.split(".");
-        const uid = Utils.base64DecodeStringUrlSafe(homeAccountId[0]);
-        const utid = Utils.base64DecodeStringUrlSafe(homeAccountId[1]);
-
-        // TODO: domain_req and login_req are not needed according to eSTS team
-        ssoParam[SSOTypes.LOGIN_REQ] = uid;
-        ssoParam[SSOTypes.DOMAIN_REQ] = utid;
-
-        if (utid === Constants.consumersUtid) {
-            ssoParam[SSOTypes.DOMAIN_HINT] = SSOTypes.CONSUMERS;
-        }
-        else {
-            ssoParam[SSOTypes.DOMAIN_HINT] = SSOTypes.ORGANIZATIONS;
-        }
-        break;
-      }
-      case SSOTypes.LOGIN_REQ: {
-        ssoParam[SSOTypes.LOGIN_REQ] = ssoData;
-        break;
-      }
-      case SSOTypes.DOMAIN_REQ: {
-        ssoParam[SSOTypes.DOMAIN_REQ] = ssoData;
-        break;
-      }
-    }
-
-    return ssoParam;
-  }
-
-  /**
-   * Utility to generate a QueryParameterString from a Key-Value mapping of extraQueryParameters passed
-   * @param extraQueryParameters
-   */
-  static generateQueryParametersString(queryParameters: QPDict): string {
-    let paramsString: string = null;
-
-    if (queryParameters) {
-      Object.keys(queryParameters).forEach((key: string) => {
-        if (paramsString == null) {
-          paramsString = `${key}=${encodeURIComponent(queryParameters[key])}`;
-        }
-        else {
-          paramsString += `&${key}=${encodeURIComponent(queryParameters[key])}`;
-        }
-     });
-    }
-
-    return paramsString;
-  }
-
-  /**
-   * Check to see if there are SSO params set in the Request
-   * @param request
-   */
-  static isSSOParam(request: AuthenticationParameters) {
-      return request && (request.account || request.sid || request.loginHint);
-  }
-
-  //#endregion
-
 }
